Add createRoom view and updateRoom view in group app

+++ techsync/group/views.py
@@ -1,14 +1,27 @@
 from django.shortcuts import render,redirect
 from .models import Room, Topic
 from .forms import RoomForm
+from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 def home(request):
-    rooms = Room.objects.all()
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) |
+        Q(description__icontains=q)
+    )
+    topics = Topic.objects.all()
+    room_count = rooms.count()
     context = {
         'rooms': rooms,
-        # 'topics': topics,
+        'topics': topics,
+        'room_count': room_count,
     }
-    return render(request, 'group/home.html',context)
+    return render(request, 'group/home.html', context)
 
+
+@login_required(login_url='login')
 def room(request, pk):
     room = Room.objects.get(id=pk)
     participants = room.participants.all()
@@ -20,6 +33,8 @@ def room(request, pk):
     }
     return render(request, 'group/room.html', context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -34,9 +49,13 @@ def createRoom(request):
     return render(request, 'group/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to view this page.')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -50,8 +69,12 @@ def updateRoom(request, pk):
         }
     return render(request, 'group/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to view this page.')
     if request.method == 'POST':
         room.delete()
         return redirect('home-group')
